Don't throw on 0 backoff in clipVelocity
This throw was an artifact from a now-fixed bug, AFAIK 0 backoff is not universally a problem, and can happen when e.g. running into a ceiling angled away from you. This needs more investigation to be sure, as the velocity clipping seems to behave strangely when doing so (but doesn't fail, so removing the throw here is a fine first step).